userinfo: Show userlog embed also
Also: move things around a bit
